Moved function to new unit Game
