+++ src/Game/Actor/CMakeLists.txt
@@ -1,4 +1,6 @@
 target_sources(uking PRIVATE
+  actDragon.cpp
+  actDragon.h
   actWeapon.cpp
   actWeapon.h
 )
