+++ src/Game/Actor/actDragon.h
@@ -1,0 +1,5 @@
+#include <math/seadVector.h>
+
+namespace ksys::act {
+    bool getDragonItemDropPosition(sead::Vector3f* target_pos, const sead::Vector3f& current_pos);
+}
